+++ src/ifpa_api/__init__.py
@@ -43,9 +43,17 @@ from ifpa_api.core.exceptions import (
     SeriesPlayerNotFoundError,
     TournamentNotLeagueError,
 )
-from ifpa_api.models.common import RankingSystem, ResultType, TimePeriod, TournamentType
+from ifpa_api.models.common import (
+    MajorTournament,
+    RankingSystem,
+    ResultType,
+    StatsRankType,
+    SystemCode,
+    TimePeriod,
+    TournamentType,
+)
 
-__version__ = "0.3.0"
+__version__ = "0.4.0"
 
 __all__ = [
     # Main client
@@ -55,6 +63,9 @@ __all__ = [
     "RankingSystem",
     "ResultType",
     "TournamentType",
+    "StatsRankType",
+    "SystemCode",
+    "MajorTournament",
     # Exceptions
     "IfpaError",
     "IfpaApiError",